apply .net standard 2.0
