--- Commerble.Postal.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26403.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Commerble.Postal", "Commerble.Postal\Commerble.Postal.csproj", "{124CB3E3-9221-461F-9C51-AC40B47BC7D3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Commerble.Postal.Tests", "Commerble.Postal.Tests\Commerble.Postal.Tests.csproj", "{051564C5-ECDE-48EC-A930-FE9A274CA191}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{124CB3E3-9221-461F-9C51-AC40B47BC7D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{124CB3E3-9221-461F-9C51-AC40B47BC7D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{124CB3E3-9221-461F-9C51-AC40B47BC7D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{124CB3E3-9221-461F-9C51-AC40B47BC7D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{051564C5-ECDE-48EC-A930-FE9A274CA191}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{051564C5-ECDE-48EC-A930-FE9A274CA191}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{051564C5-ECDE-48EC-A930-FE9A274CA191}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{051564C5-ECDE-48EC-A930-FE9A274CA191}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
