--- Commerble.Postal/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommandLineParser" version="1.9.71" targetFramework="net452" />
-  <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net46" />
-  <package id="RazorTemplates" version="1.2.1" targetFramework="net452" />
-</packages>
