+++ src/Commerble.Postal/Commerble.Postal.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <Company>Commerble</Company>
+    <Authors>Commerble</Authors>
+    <PackageProjectUrl>https://github.com/commerble/postal</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/commerble/postal</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+  </PropertyGroup>
+
+</Project>
