+++ src/Commerble.Postal/Commerble.Postal.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29009.5
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Commerble.Postal", "Commerble.Postal.csproj", "{017C5D67-0ABA-4163-BB66-91CF7404A250}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Commerble.Postal.Tests", "..\Commerble.Postal.Tests\Commerble.Postal.Tests.csproj", "{BF8BAE33-5BF5-4B37-A8E2-9003999481DF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Commerble.Postal.ConsoleApp", "..\Commerble.Postal.ConsoleApp\Commerble.Postal.ConsoleApp.csproj", "{47841AD1-6C2B-4BEB-B76E-C8063B6FC3B7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{017C5D67-0ABA-4163-BB66-91CF7404A250}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{017C5D67-0ABA-4163-BB66-91CF7404A250}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{017C5D67-0ABA-4163-BB66-91CF7404A250}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{017C5D67-0ABA-4163-BB66-91CF7404A250}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF8BAE33-5BF5-4B37-A8E2-9003999481DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF8BAE33-5BF5-4B37-A8E2-9003999481DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF8BAE33-5BF5-4B37-A8E2-9003999481DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF8BAE33-5BF5-4B37-A8E2-9003999481DF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47841AD1-6C2B-4BEB-B76E-C8063B6FC3B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47841AD1-6C2B-4BEB-B76E-C8063B6FC3B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47841AD1-6C2B-4BEB-B76E-C8063B6FC3B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47841AD1-6C2B-4BEB-B76E-C8063B6FC3B7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2C5372C4-C3A1-4073-82A6-6C03C4D01ED2}
+	EndGlobalSection
+EndGlobal
